+++ changelog.md
@@ -13,12 +13,17 @@
 * views/partials/menubar.php -CL
 * veiws/partials/footer.php -CL
 --------Feb 8, 2018--------
+* 3 models - QL
+* Accessories.php, Categories.php, Presets.php
+* 3 controllers - QL
+* AccessoriesCtr.php, CategoriesCtr.php, PresetsCtr.php
 * csv folder - QL
 * 3 csv files - QL
 * Accessories.csv, Categories.csv, Presets.csv - QL
 * images folder - RT
 * 31 pictures - RT
 * weapon folder and weapon images - RT
+* about.php controllers and about_message view - RT
 
 ### Updated components
 --------Feb 9,2018